type fix in new parser

+++ Engine/csound_orc_compile.c
@@ -24,6 +24,7 @@
 */
 
 #include "csoundCore.h"
+#include "parse_param.h"
 #include "csound_orc.h"
 #include <math.h>
 #include <ctype.h>
@@ -1927,10 +1928,10 @@ char argtyp2(CSOUND *csound, char *s)
 }
 
 /* For diagnostics map file name or macro name to an index */
-int file_to_int(CSOUND *csound, const char *name)
+uint8_t file_to_int(CSOUND *csound, const char *name)
 {
     extern char *strdup(const char *);
-    int n = 0;
+    uint8_t n = 0;
     char **filedir = csound->filedir;
     while (filedir[n] && n<63) {        /* Do we have it already? */
       if (strcmp(filedir[n], name)==0) return n; /* yes */
